pyproject.toml
[project]
name = "ansify"
version = "0.0.0"
dependencies = ["pillow (>=11.2.1,<12.0.0)", "numpy (>=2.2.6,<3.0.0)", "matplotlib (>=3.10.3,<4.0.0)"]
requires-python = ">=3.12"

[project.scripts]
ansify = "ansify:main"
